refactor: update brc20 page rows
