+++ components/TruncatingTooltip.tsx
@@ -3,11 +3,12 @@ import { Tooltip, TooltipContent, TooltipTrigger } from "./Tooltip";
 
 const TruncatingTooltip = (props: { num: string }) => {
   const truncated = truncateAmount(props.num);
+  console.log(truncated);
   return (
     <Tooltip>
       <TooltipTrigger>
         {truncated}
-        {props.num === truncated || <>&hellip;</>}
+        {truncated.includes(".") && props.num !== truncated && <>&hellip;</>}
       </TooltipTrigger>
       <TooltipContent variant="light">{props.num}</TooltipContent>
     </Tooltip>
