+++ lib/utils.ts
@@ -103,7 +103,8 @@ export function humanReadableCount(number: number, precision: number): string {
 
 export function truncateAmount(num: string, maxDecimals = 2) {
   const [whole, decimals] = num.split(".");
-  // return only `whole` if `decimals` is all zeroes
+
+  if (decimals && !decimals.replace(/0/g, "")) return whole; // return only `whole` if `decimals` is all zeroes
 
   if (!decimals || decimals.length <= maxDecimals) return num;
   return `${whole}.${decimals.substring(0, maxDecimals)}`;
